requirements.txt
atpbar>=1.1.4
dpkt>=1.7
netaddr>=0.7.19
numpy>=1.24.3
pandas>=2.0.2
progressbar33>=2.4
randmac>=0.0.2
scapy>=2.5.0
